+++ src/soloud/src/audiosource/wav/soloud_wav.cpp
@@ -87,7 +87,7 @@ namespace SoLoud
 			long samples = mScratchSize / mChannels;
 			if (samples > samples_to_discard)
 				samples = samples_to_discard;
-			getAudio(mScratch, samples, samples);
+			getAudio(mScratch, (unsigned int)samples, (unsigned int)samples);
 			samples_to_discard -= samples;
 		}
 		int pos = (int)floor(mBaseSamplerate * aSeconds);
